Provides new FContactListener interface for communicating contact events.
The old style, of providing particular function names in a Processing sketch,
still works fine. But this permits more advanced programmers to encapsulate
their contact handling. It also permits Python Mode to use Fisica, since
a Jython class can subclass a Java class, but it's not feasible for Fisica
to find method names implemented in the Python sketch code so easily.